Refactor: Add clock component for time updating
