+++ src/components/Date/LocaleTimer.js
@@ -1,22 +1,14 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import FormattedDate from './FormattedDate';
+import { getFullLocaleDate } from 'utils/date';
 
 LocaleTimer.propTypes = {
   date: PropTypes.instanceOf(Date),
   locale: PropTypes.string,
-  handleDate: PropTypes.func,
 };
 
-function LocaleTimer({ date, locale, handleDate }) {
-  useEffect(() => {
-    if (handleDate) {
-      const timerId = setInterval(() => handleDate(), 1000);
-      return () => clearInterval(timerId);
-    }
-  }, [handleDate]);
-
-  return <FormattedDate date={date} locale={locale}></FormattedDate>;
+function LocaleTimer({ date, locale }) {
+  return <span>{getFullLocaleDate(date, locale)}</span>;
 }
 
 export default LocaleTimer;
